+++ frontend/src/components/app.js
@@ -6,13 +6,11 @@ import NavBarContainer from "./nav/navbar_container";
 import MainPage from "./main/main_page";
 import LoginFormContainer from "./session/login_form_container";
 import SignupFormContainer from "./session/signup_form_container";
-// import ApiCalendar from "./calendar/calendar";
-import GoogleAuth from "./session/google_auth";
+import GoogleLogin from "./session/google_login";
 
 const App = () => (
   <section>
     <NavBarContainer />
-    <GoogleAuth/>
     <Switch>
       <AuthRoute exact path="/" component={MainPage} />
       <AuthRoute exact path="/login" component={LoginFormContainer} />
